Move testdata to top-level package.
- remove last remnants of dart support.
- move Bazel.java to new package.
- rename okhttp3.okhttp repository.
