+++ go/screenshotter/screenshotter_test.go
@@ -38,7 +38,7 @@ func TestMain(m *testing.M) {
 		log.Fatal(err)
 	}
 
-	dir, err := bazel.RunfilesPath()
+	dir, err := bazel.Runfile("testdata/")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -49,7 +49,7 @@ func TestMain(m *testing.M) {
 		log.Fatal(http.ListenAndServe(fmt.Sprintf(":%d", port), http.FileServer(http.Dir(dir))))
 	}()
 
-	testpage = fmt.Sprintf("http://localhost:%d/go/launcher/proxy/testdata/testpage.html", port)
+	testpage = fmt.Sprintf("http://localhost:%d/testpage.html", port)
 
 	os.Exit(m.Run())
 }
