+++ javatests/com/google/testing/web/screenshotter/BUILD
@@ -27,7 +27,7 @@ java_web_test_suite(
         "//browsers:chromium-native",
         "//browsers:firefox-native",
     ],
-    data = ["//go/launcher/proxy/testdata"],
+    data = ["//testdata"],
     java_test_tags = [
         "manual",
         "noci",
@@ -35,7 +35,7 @@ java_web_test_suite(
     local = True,
     tags = ["noci"],
     deps = [
-        "//java/com/google/testing/util",
+        "//java/com/google/testing/bazel",
         "//java/com/google/testing/web",
         "//java/com/google/testing/web/screenshotter",
         "@junit",
