+++ web/BUILD
@@ -53,15 +53,6 @@ filegroup(
 )
 
 filegroup(
-    name = "dart",
-    srcs = ["dart.bzl"],
-    data = [
-        "//web/internal:wrap_web_test_suite",
-        # should depend on dart build rules.
-    ],
-)
-
-filegroup(
     name = "go",
     srcs = ["go.bzl"],
     data = [
