--- web/dart.bzl
@@ -1,42 +1,0 @@
-# Copyright 2016 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Web Test rules for Dart."""
-
-load("@io_bazel_rules_dart//dart/build_rules:vm.bzl", "dart_vm_test")
-load("//web/internal:wrap_web_test_suite.bzl", "wrap_web_test_suite")
-
-
-def dart_web_test_suite(name,
-                        dart_test_tags=("manual",),
-                        script_file=None,
-                        **kwargs):
-  """Defines a test_suite of web_test targets that wrap a dart_vm_test target.
-
-  Args:
-    name: The base name of the test.
-    dart_test_tags: A list of test tag strings to use for the dart_test target.
-      Defaults to ['manual'].
-    script_file: The name of the source file that is the main entry point of the
-      application. (default: computed from name)
-    **kwargs: Arguments for wrapped_web_test_suite
-  """
-  if script_file == None:
-    script_file = "test/" + name + ".dart"
-
-  wrap_web_test_suite(
-      name=name,
-      rule=dart_vm_test,
-      script_file=script_file,
-      wrapped_test_tags=dart_test_tags,
-      **kwargs)
